Updated form for the editable form
