+++ apps/dashboard/src/components/forms/notifications/form-telegram.tsx
@@ -15,21 +15,19 @@ import {
   FormCardSeparator,
 } from "@/components/forms/form-card";
 import { useFormSheetDirty } from "@/components/forms/form-sheet";
-import { useTRPC } from "@/lib/trpc/client";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Button } from "@openstatus/ui/components/ui/button";
 import { Form } from "@openstatus/ui/components/ui/form";
 import { Input } from "@openstatus/ui/components/ui/input";
 import { Label } from "@openstatus/ui/components/ui/label";
 import { cn } from "@openstatus/ui/lib/utils";
-import { useMutation, useQuery } from "@tanstack/react-query";
 import { isTRPCClientError } from "@trpc/client";
 import React, { useTransition } from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
 import { z } from "zod";
+import { TelegramConnectionFlow } from "../components/telegram-connection-flow";
+import { TelegramFormActions } from "../components/telegram-form-actions";
 import { TelegramManualInput } from "../components/telegram-manual-input";
-import { TelegramQRConnection } from "../components/telegram-qr-connection";
 
 const schema = z.object({
   name: z.string(),
@@ -68,88 +66,18 @@ export function FormTelegram({
   });
   const [isPending, startTransition] = useTransition();
   const { setIsDirty } = useFormSheetDirty();
-  const trpc = useTRPC();
-  const sendTestMutation = useMutation(
-    trpc.notification.sendTest.mutationOptions(),
-  );
 
   const formIsDirty = form.formState.isDirty;
   React.useEffect(() => {
     setIsDirty(formIsDirty);
   }, [formIsDirty, setIsDirty]);
 
-  // Create Telegram Token
-  const { data: tokenData, isLoading: isTokenLoading } = useQuery({
-    ...trpc.notification.createTelegramToken.queryOptions(),
-    refetchOnWindowFocus: false,
-  });
+  // Check if we're editing an existing notification (has chatID) or creating a new one
+  const isEditMode = React.useMemo(() => {
+    return Boolean(defaultValues?.data?.chatId);
+  }, [defaultValues]);
 
   const [mode, setMode] = React.useState<"qr" | "manual" | null>(null);
-  const [flowStep, setFlowStep] = React.useState<"private" | "group">(
-    "private",
-  );
-  const [privateChatId, setPrivateChatId] = React.useState<string | null>(null);
-  const [sessionStartTime, setSessionStartTime] = React.useState<number | null>(
-    null,
-  );
-
-  // Set session start time when entering QR mode
-  React.useEffect(() => {
-    if (mode === "qr") {
-      setSessionStartTime(Math.floor(Date.now() / 1000));
-    } else if (mode === null) {
-      setSessionStartTime(null);
-    }
-  }, [mode]);
-
-  // Cleanup: Reset UI state when component unmounts (e.g., on discard)
-  React.useEffect(() => {
-    return () => {
-      // This runs when component unmounts
-      setMode(null);
-      setFlowStep("private");
-      setPrivateChatId(null);
-      setSessionStartTime(null);
-    };
-  }, []);
-
-  // Start polling for updates
-  const { data: updates } = useQuery({
-    ...trpc.notification.getTelegramUpdates.queryOptions({
-      privateChatId:
-        flowStep === "group" ? privateChatId ?? undefined : undefined,
-      since: sessionStartTime ?? undefined,
-    }),
-    enabled:
-      !!tokenData?.token && !form.getValues("data.chatId") && mode === "qr",
-    refetchInterval: 5000,
-  });
-
-  React.useEffect(() => {
-    if (updates && updates.length > 0) {
-      const lastUpdate = updates[updates.length - 1];
-
-      // Phase 1: Private chat ID received
-      if (lastUpdate.chatType === "private" && flowStep === "private") {
-        setPrivateChatId(lastUpdate.chatId);
-        setFlowStep("group");
-        toast.success(
-          `Connected to ${lastUpdate.user.first_name}'s account. Now add the bot to your group.`,
-        );
-      }
-      // Phase 2: Group chat ID received
-      else if (lastUpdate.chatType === "group" && flowStep === "group") {
-        startTransition(() => {
-          form.setValue("data.chatId", lastUpdate.chatId, {
-            shouldDirty: true,
-          });
-          toast.success(
-            `Connected to group "${lastUpdate.chatTitle || "Unknown"}"`,
-          );
-        });
-      }
-    }
-  }, [updates, form, flowStep]);
 
   function submitAction(values: FormValues) {
     if (isPending) return;
@@ -171,42 +99,7 @@ export function FormTelegram({
 
         // Reset UI state after successful submission
         setMode(null);
-        setFlowStep("private");
-        setPrivateChatId(null);
         form.reset();
-      } catch (error) {
-        console.error(error);
-      }
-    });
-  }
-
-  function testAction() {
-    if (isPending) return;
-
-    startTransition(async () => {
-      try {
-        const provider = form.getValues("provider");
-        const data = form.getValues("data");
-        const promise = sendTestMutation.mutateAsync({
-          provider,
-          data: {
-            telegram: { chatId: data.chatId },
-          },
-        });
-        toast.promise(promise, {
-          loading: "Sending test...",
-          success: "Test sent",
-          error: (error) => {
-            if (isTRPCClientError(error)) {
-              return error.message;
-            }
-            if (error instanceof Error) {
-              return error.message;
-            }
-            return "Failed to send test";
-          },
-        });
-        await promise;
       } catch (error) {
         console.error(error);
       }
@@ -238,64 +131,19 @@ export function FormTelegram({
             )}
           />
           <div className="flex flex-col gap-4">
-            <div className="grid grid-cols-[1fr_auto_1fr] items-center gap-4">
-              <Button
-                type="button"
-                variant={mode === "qr" ? "default" : "outline"}
-                size="sm"
-                onClick={() => setMode("qr")}
-                className="cursor-pointer w-full"
-              >
-                Connect with QR
-              </Button>
-              <div className="flex items-center gap-2 min-w-[40px]">
-                <div className="h-px bg-border flex-1" />
-                <span className="text-[10px] text-muted-foreground font-bold uppercase tracking-wider">
-                  Or
-                </span>
-                <div className="h-px bg-border flex-1" />
-              </div>
-              <Button
-                type="button"
-                variant={mode === "manual" ? "default" : "outline"}
-                size="sm"
-                onClick={() => setMode("manual")}
-                className="cursor-pointer w-full"
-              >
-                Enter ChatID manually
-              </Button>
-            </div>
-
-            {mode === "manual" && <TelegramManualInput form={form} />}
-            {mode === "qr" && (
-              <TelegramQRConnection
+            {isEditMode ? (
+              // Edit mode: Show editable chatID input only
+              <TelegramManualInput form={form} />
+            ) : (
+              // Create mode: Show QR/manual connection flow
+              <TelegramConnectionFlow
                 form={form}
-                token={tokenData?.token}
-                isLoading={isTokenLoading}
-                isPolling={
-                  !!tokenData?.token &&
-                  !form.watch("data.chatId") &&
-                  mode === "qr"
-                }
-                flowStep={flowStep}
-                privateChatId={privateChatId}
-                onReset={() => {
-                  form.setValue("data.chatId", "", { shouldDirty: true });
-                  setSessionStartTime(Math.floor(Date.now() / 1000));
-                }}
+                mode={mode}
+                onModeChange={setMode}
               />
             )}
           </div>
-          <div>
-            <Button
-              variant="outline"
-              size="sm"
-              type="button"
-              onClick={testAction}
-            >
-              Send Test
-            </Button>
-          </div>
+          <TelegramFormActions form={form} isPending={isPending} />
         </FormCardContent>
         <FormCardSeparator />
         <FormCardContent>
